+++ BBStats/src/main/java/com/uep/wap/repository/StatisticianRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import com.uep.wap.model.Statistician;
 
 @Repository
-public interface StatisticianRepository extends CrudRepository<Statistician,Integer> {
+public interface StatisticianRepository extends CrudRepository<Statistician,Long> {
 
 }
